+++ src/actions-util.ts
@@ -57,13 +57,6 @@ export function getTemporaryDirectory(): string {
     : getRequiredEnvParam("RUNNER_TEMP");
 }
 
-export function getToolCacheDirectory(): string {
-  const value = process.env["CODEQL_ACTION_TOOL_CACHE"];
-  return value !== undefined && value !== ""
-    ? value
-    : getRequiredEnvParam("RUNNER_TOOL_CACHE");
-}
-
 /**
  * Gets the SHA of the commit that is currently checked out.
  */
@@ -674,12 +667,7 @@ export async function createStatusReportBase(
   }
   const runnerOs = getRequiredEnvParam("RUNNER_OS");
   const codeQlCliVersion = getCachedCodeQlVersion();
-
-  // If running locally then the GITHUB_ACTION_REF cannot be trusted as it may be for the previous action
-  // See https://github.com/actions/runner/issues/803
-  const actionRef = isRunningLocalAction()
-    ? undefined
-    : process.env["GITHUB_ACTION_REF"];
+  const actionRef = process.env["GITHUB_ACTION_REF"];
 
   const statusReport: StatusReportBase = {
     workflow_run_id: workflowRunID,