+++ src/codeql.ts
@@ -3,20 +3,25 @@ import { OutgoingHttpHeaders } from "http";
 import * as path from "path";
 
 import * as toolrunner from "@actions/exec/lib/toolrunner";
+import * as toolcache from "@actions/tool-cache";
 import { default as deepEqual } from "fast-deep-equal";
 import { default as queryString } from "query-string";
 import * as semver from "semver";
+import { v4 as uuidV4 } from "uuid";
 
-import { isRunningLocalAction, getRelativeScriptPath } from "./actions-util";
+import { getRelativeScriptPath, isRunningLocalAction } from "./actions-util";
 import * as api from "./api-client";
 import { Config } from "./config-utils";
 import * as defaults from "./defaults.json"; // Referenced from codeql-action-sync-tool!
 import { errorMatchers } from "./error-matcher";
-import { FeatureFlags, FeatureFlag } from "./feature-flags";
+import { FeatureFlag, FeatureFlags } from "./feature-flags";
 import { isTracedLanguage, Language } from "./languages";
 import { Logger } from "./logging";
-import * as toolcache from "./toolcache";
 import { toolrunnerErrorCatcher } from "./toolrunner-error-catcher";
+import {
+  getTrapCachingExtractorConfigArgs,
+  getTrapCachingExtractorConfigArgsForLang,
+} from "./trap-caching";
 import * as util from "./util";
 import { isGoodVersion } from "./util";
 
@@ -95,7 +100,11 @@ export interface CodeQL {
    * Extract code for a scanned language using 'codeql database trace-command'
    * and running the language extractor.
    */
-  extractScannedLanguage(database: string, language: Language): Promise<void>;
+  extractScannedLanguage(
+    config: Config,
+    language: Language,
+    featureFlags: FeatureFlags
+  ): Promise<void>;
   /**
    * Finalize a database using 'codeql database finalize'.
    */
@@ -108,6 +117,10 @@ export interface CodeQL {
    * Run 'codeql resolve languages'.
    */
   resolveLanguages(): Promise<ResolveLanguagesOutput>;
+  /**
+   * Run 'codeql resolve languages' with '--format=betterjson'.
+   */
+  betterResolveLanguages(): Promise<BetterResolveLanguagesOutput>;
   /**
    * Run 'codeql resolve queries'.
    */
@@ -152,6 +165,7 @@ export interface CodeQL {
     sarifFile: string,
     addSnippetsFlag: string,
     threadsFlag: string,
+    verbosityFlag: string | undefined,
     automationDetailsId: string | undefined
   ): Promise<string>;
   /**
@@ -162,6 +176,17 @@ export interface CodeQL {
 
 export interface ResolveLanguagesOutput {
   [language: string]: [string];
+}
+
+export interface BetterResolveLanguagesOutput {
+  extractors: {
+    [language: string]: [
+      {
+        extractor_root: string;
+        extractor_options?: any;
+      }
+    ];
+  };
 }
 
 export interface ResolveQueriesOutput {
@@ -241,6 +266,12 @@ export const CODEQL_VERSION_NEW_TRACING = "2.7.0";
  * some of their files being greater than MAX_PATH (260 characters).
  */
 export const CODEQL_VERSION_ML_POWERED_QUERIES_WINDOWS = "2.9.0";
+
+/**
+ * Previous versions had the option already, but were missing the
+ * --extractor-options-verbosity that we need.
+ */
+export const CODEQL_VERSION_BETTER_RESOLVE_LANGUAGES = "2.10.3";
 
 function getCodeQLBundleName(): string {
   let platform: string;
@@ -381,7 +412,6 @@ async function getCodeQLBundleDownloadURL(
  * @param codeqlURL
  * @param apiDetails
  * @param tempDir
- * @param toolCacheDir
  * @param variant
  * @param logger
  * @param checkVersion Whether to check that CodeQL CLI meets the minimum
@@ -392,7 +422,6 @@ export async function setupCodeQL(
   codeqlURL: string | undefined,
   apiDetails: api.GitHubApiDetails,
   tempDir: string,
-  toolCacheDir: string,
   variant: util.GitHubVariant,
   logger: Logger,
   checkVersion: boolean
@@ -407,7 +436,7 @@ export async function setupCodeQL(
     let codeqlFolder: string;
     let codeqlURLVersion: string;
     if (codeqlURL && !codeqlURL.startsWith("http")) {
-      codeqlFolder = await toolcache.extractTar(codeqlURL, tempDir, logger);
+      codeqlFolder = await toolcache.extractTar(codeqlURL);
       codeqlURLVersion = "local";
     } else {
       codeqlURLVersion = getCodeQLURLVersion(
@@ -416,29 +445,15 @@ export async function setupCodeQL(
       const codeqlURLSemVer = convertToSemVer(codeqlURLVersion, logger);
 
       // If we find the specified version, we always use that.
-      codeqlFolder = toolcache.find(
-        "CodeQL",
-        codeqlURLSemVer,
-        toolCacheDir,
-        logger
-      );
+      codeqlFolder = toolcache.find("CodeQL", codeqlURLSemVer);
 
       // If we don't find the requested version, in some cases we may allow a
       // different version to save download time if the version hasn't been
       // specified explicitly (in which case we always honor it).
       if (!codeqlFolder && !codeqlURL && !forceLatest) {
-        const codeqlVersions = toolcache.findAllVersions(
-          "CodeQL",
-          toolCacheDir,
-          logger
-        );
+        const codeqlVersions = toolcache.findAllVersions("CodeQL");
         if (codeqlVersions.length === 1 && isGoodVersion(codeqlVersions[0])) {
-          const tmpCodeqlFolder = toolcache.find(
-            "CodeQL",
-            codeqlVersions[0],
-            toolCacheDir,
-            logger
-          );
+          const tmpCodeqlFolder = toolcache.find("CodeQL", codeqlVersions[0]);
           if (fs.existsSync(path.join(tmpCodeqlFolder, "pinned-version"))) {
             logger.debug(
               `CodeQL in cache overriding the default ${CODEQL_BUNDLE_VERSION}`
@@ -480,24 +495,25 @@ export async function setupCodeQL(
         logger.info(
           `Downloading CodeQL tools from ${codeqlURL}. This may take a while.`
         );
+
+        const dest = path.join(tempDir, uuidV4());
+        const finalHeaders = Object.assign(
+          { "User-Agent": "CodeQL Action" },
+          headers
+        );
         const codeqlPath = await toolcache.downloadTool(
           codeqlURL,
-          tempDir,
-          headers
+          dest,
+          undefined,
+          finalHeaders
         );
         logger.debug(`CodeQL bundle download to ${codeqlPath} complete.`);
 
-        const codeqlExtracted = await toolcache.extractTar(
-          codeqlPath,
-          tempDir,
-          logger
-        );
+        const codeqlExtracted = await toolcache.extractTar(codeqlPath);
         codeqlFolder = await toolcache.cacheDir(
           codeqlExtracted,
           "CodeQL",
-          codeqlURLSemVer,
-          toolCacheDir,
-          logger
+          codeqlURLSemVer
         );
       }
     }
@@ -594,6 +610,10 @@ export function setCodeQL(partialCodeql: Partial<CodeQL>): CodeQL {
     ),
     finalizeDatabase: resolveFunction(partialCodeql, "finalizeDatabase"),
     resolveLanguages: resolveFunction(partialCodeql, "resolveLanguages"),
+    betterResolveLanguages: resolveFunction(
+      partialCodeql,
+      "betterResolveLanguages"
+    ),
     resolveQueries: resolveFunction(partialCodeql, "resolveQueries"),
     packDownload: resolveFunction(partialCodeql, "packDownload"),
     databaseCleanup: resolveFunction(partialCodeql, "databaseCleanup"),
@@ -630,8 +650,10 @@ export function getCachedCodeQL(): CodeQL {
  * a non-existent placeholder codeql command, so tests that use this function
  * should also stub the toolrunner.ToolRunner constructor.
  */
-export async function getCodeQLForTesting(): Promise<CodeQL> {
-  return getCodeQLForCmd("codeql-for-testing", false);
+export async function getCodeQLForTesting(
+  cmd = "codeql-for-testing"
+): Promise<CodeQL> {
+  return getCodeQLForCmd(cmd, false);
 }
 
 /**
@@ -737,6 +759,7 @@ async function getCodeQLForCmd(
       );
       if (config.languages.filter(isTracedLanguage).length > 0) {
         extraArgs.push("--begin-tracing");
+        extraArgs.push(...(await getTrapCachingExtractorConfigArgs(config)));
         if (processName !== undefined) {
           extraArgs.push(`--trace-process-name=${processName}`);
         } else {
@@ -787,9 +810,28 @@ async function getCodeQLForCmd(
         "-Dmaven.wagon.http.pool=false",
       ].join(" ");
 
+      // On macOS, System Integrity Protection (SIP) typically interferes with
+      // CodeQL build tracing of protected binaries.
+      // The usual workaround is to prefix `$CODEQL_RUNNER` to build commands:
+      // `$CODEQL_RUNNER` (not to be confused with the deprecated CodeQL Runner tool)
+      // points to a simple wrapper binary included with the CLI, and the extra layer of
+      // process indirection helps the tracer bypass SIP.
+
+      // The above SIP workaround is *not* needed here.
+      // At the `autobuild` step in the Actions workflow, we assume the `init` step
+      // has successfully run, and will have exported `DYLD_INSERT_LIBRARIES`
+      // into the environment of subsequent steps, to activate the tracer.
+      // When `DYLD_INSERT_LIBRARIES` is set in the environment for a step,
+      // the Actions runtime introduces its own workaround for SIP
+      // (https://github.com/actions/runner/pull/416).
       await runTool(autobuildCmd);
     },
-    async extractScannedLanguage(databasePath: string, language: Language) {
+    async extractScannedLanguage(
+      config: Config,
+      language: Language,
+      featureFlags: FeatureFlags
+    ) {
+      const databasePath = util.getCodeQLDatabasePath(config, language);
       // Get extractor location
       let extractorPath = "";
       await new toolrunner.ToolRunner(
@@ -821,6 +863,16 @@ async function getCodeQLForCmd(
         "tools",
         `autobuild${ext}`
       );
+      const extraArgs: string[] = [];
+      if (
+        await util.codeQlVersionAbove(this, CODEQL_VERSION_LUA_TRACER_CONFIG)
+      ) {
+        if (await featureFlags.getValue(FeatureFlag.LuaTracerConfigEnabled)) {
+          extraArgs.push("--internal-use-lua-tracing");
+        } else {
+          extraArgs.push("--no-internal-use-lua-tracing");
+        }
+      }
 
       // Run trace command
       await toolrunnerErrorCatcher(
@@ -828,6 +880,8 @@ async function getCodeQLForCmd(
         [
           "database",
           "trace-command",
+          ...extraArgs,
+          ...(await getTrapCachingExtractorConfigArgsForLang(config, language)),
           ...getExtraOptionsFromEnv(["database", "trace-command"]),
           databasePath,
           "--",
@@ -867,6 +921,24 @@ async function getCodeQLForCmd(
       } catch (e) {
         throw new Error(
           `Unexpected output from codeql resolve languages: ${e}`
+        );
+      }
+    },
+    async betterResolveLanguages() {
+      const codeqlArgs = [
+        "resolve",
+        "languages",
+        "--format=betterjson",
+        "--extractor-options-verbosity=4",
+        ...getExtraOptionsFromEnv(["resolve", "languages"]),
+      ];
+      const output = await runTool(cmd, codeqlArgs);
+
+      try {
+        return JSON.parse(output);
+      } catch (e) {
+        throw new Error(
+          `Unexpected output from codeql resolve languages with --format=betterjson: ${e}`
         );
       }
     },
@@ -921,6 +993,7 @@ async function getCodeQLForCmd(
       sarifFile: string,
       addSnippetsFlag: string,
       threadsFlag: string,
+      verbosityFlag: string,
       automationDetailsId: string | undefined
     ): Promise<string> {
       const codeqlArgs = [
@@ -928,7 +1001,7 @@ async function getCodeQLForCmd(
         "interpret-results",
         threadsFlag,
         "--format=sarif-latest",
-        "-v",
+        verbosityFlag,
         `--output=${sarifFile}`,
         addSnippetsFlag,
         ...getExtraOptionsFromEnv(["database", "interpret-results"]),