+++ src/config-utils.test.ts
@@ -86,10 +86,10 @@ test("load empty config", async (t) => {
       undefined,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -107,10 +107,10 @@ test("load empty config", async (t) => {
         undefined,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         codeQL,
         tmpDir,
@@ -153,10 +153,10 @@ test("loading config saves config", async (t) => {
       undefined,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -188,10 +188,10 @@ test("load input outside of workspace", async (t) => {
         "../input",
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -227,10 +227,10 @@ test("load non-local input with invalid repo syntax", async (t) => {
         configFile,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -267,10 +267,10 @@ test("load non-existent input", async (t) => {
         configFile,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -348,7 +348,6 @@ test("load non-empty input", async (t) => {
         paths: ["c/d"],
       },
       tempDir: tmpDir,
-      toolCacheDir: tmpDir,
       codeQLCmd: codeQL.getPath(),
       gitHubVersion,
       dbLocation: path.resolve(tmpDir, "codeql_databases"),
@@ -357,6 +356,7 @@ test("load non-empty input", async (t) => {
       debugArtifactName: "my-artifact",
       debugDatabaseName: "my-db",
       injectedMlQueries: false,
+      trapCaches: {},
     };
 
     const languages = "javascript";
@@ -369,10 +369,10 @@ test("load non-empty input", async (t) => {
       configFilePath,
       undefined,
       false,
+      false,
       "my-artifact",
       "my-db",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -436,10 +436,10 @@ test("Default queries are used", async (t) => {
       configFilePath,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -511,10 +511,10 @@ test("Queries can be specified in config file", async (t) => {
       configFilePath,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -529,16 +529,21 @@ test("Queries can be specified in config file", async (t) => {
     // and once for `./foo` from the config file.
     t.deepEqual(resolveQueriesArgs.length, 2);
     t.deepEqual(resolveQueriesArgs[1].queries.length, 1);
-    t.regex(resolveQueriesArgs[1].queries[0], /.*\/foo$/);
+    t.true(resolveQueriesArgs[1].queries[0].endsWith(`${path.sep}foo`));
 
     // Now check that the end result contains the default queries and the query from config
     t.deepEqual(config.queries["javascript"].builtin.length, 1);
     t.deepEqual(config.queries["javascript"].custom.length, 1);
-    t.regex(
-      config.queries["javascript"].builtin[0],
-      /javascript-code-scanning.qls$/
+    t.true(
+      config.queries["javascript"].builtin[0].endsWith(
+        "javascript-code-scanning.qls"
+      )
     );
-    t.regex(config.queries["javascript"].custom[0].queries[0], /.*\/foo$/);
+    t.true(
+      config.queries["javascript"].custom[0].queries[0].endsWith(
+        `${path.sep}foo`
+      )
+    );
   });
 });
 
@@ -580,10 +585,10 @@ test("Queries from config file can be overridden in workflow file", async (t) =>
       configFilePath,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -598,16 +603,21 @@ test("Queries from config file can be overridden in workflow file", async (t) =>
     // but won't be called for './foo' from the config file.
     t.deepEqual(resolveQueriesArgs.length, 2);
     t.deepEqual(resolveQueriesArgs[1].queries.length, 1);
-    t.regex(resolveQueriesArgs[1].queries[0], /.*\/override$/);
+    t.true(resolveQueriesArgs[1].queries[0].endsWith(`${path.sep}override`));
 
     // Now check that the end result contains only the default queries and the override query
     t.deepEqual(config.queries["javascript"].builtin.length, 1);
     t.deepEqual(config.queries["javascript"].custom.length, 1);
-    t.regex(
-      config.queries["javascript"].builtin[0],
-      /javascript-code-scanning.qls$/
+    t.true(
+      config.queries["javascript"].builtin[0].endsWith(
+        "javascript-code-scanning.qls"
+      )
     );
-    t.regex(config.queries["javascript"].custom[0].queries[0], /.*\/override$/);
+    t.true(
+      config.queries["javascript"].custom[0].queries[0].endsWith(
+        `${path.sep}override`
+      )
+    );
   });
 });
 
@@ -647,10 +657,10 @@ test("Queries in workflow file can be used in tandem with the 'disable default q
       configFilePath,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -665,14 +675,17 @@ test("Queries in workflow file can be used in tandem with the 'disable default q
     // but won't be called for the default one since that was disabled
     t.deepEqual(resolveQueriesArgs.length, 1);
     t.deepEqual(resolveQueriesArgs[0].queries.length, 1);
-    t.regex(resolveQueriesArgs[0].queries[0], /.*\/workflow-query$/);
+    t.true(
+      resolveQueriesArgs[0].queries[0].endsWith(`${path.sep}workflow-query`)
+    );
 
     // Now check that the end result contains only the workflow query, and not the default one
     t.deepEqual(config.queries["javascript"].builtin.length, 0);
     t.deepEqual(config.queries["javascript"].custom.length, 1);
-    t.regex(
-      config.queries["javascript"].custom[0].queries[0],
-      /.*\/workflow-query$/
+    t.true(
+      config.queries["javascript"].custom[0].queries[0].endsWith(
+        `${path.sep}workflow-query`
+      )
     );
   });
 });
@@ -707,10 +720,10 @@ test("Multiple queries can be specified in workflow file, no config file require
       undefined,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -726,23 +739,26 @@ test("Multiple queries can be specified in workflow file, no config file require
     t.deepEqual(resolveQueriesArgs.length, 3);
     t.deepEqual(resolveQueriesArgs[1].queries.length, 1);
     t.deepEqual(resolveQueriesArgs[2].queries.length, 1);
-    t.regex(resolveQueriesArgs[1].queries[0], /.*\/override1$/);
-    t.regex(resolveQueriesArgs[2].queries[0], /.*\/override2$/);
+    t.true(resolveQueriesArgs[1].queries[0].endsWith(`${path.sep}override1`));
+    t.true(resolveQueriesArgs[2].queries[0].endsWith(`${path.sep}override2`));
 
     // Now check that the end result contains both the queries from the workflow, as well as the defaults
     t.deepEqual(config.queries["javascript"].builtin.length, 1);
     t.deepEqual(config.queries["javascript"].custom.length, 2);
-    t.regex(
-      config.queries["javascript"].builtin[0],
-      /javascript-code-scanning.qls$/
+    t.true(
+      config.queries["javascript"].builtin[0].endsWith(
+        "javascript-code-scanning.qls"
+      )
     );
-    t.regex(
-      config.queries["javascript"].custom[0].queries[0],
-      /.*\/override1$/
+    t.true(
+      config.queries["javascript"].custom[0].queries[0].endsWith(
+        `${path.sep}override1`
+      )
     );
-    t.regex(
-      config.queries["javascript"].custom[1].queries[0],
-      /.*\/override2$/
+    t.true(
+      config.queries["javascript"].custom[1].queries[0].endsWith(
+        `${path.sep}override2`
+      )
     );
   });
 });
@@ -788,10 +804,10 @@ test("Queries in workflow file can be added to the set of queries without overri
       configFilePath,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -807,28 +823,35 @@ test("Queries in workflow file can be added to the set of queries without overri
     // and once for './foo' from the config file
     t.deepEqual(resolveQueriesArgs.length, 4);
     t.deepEqual(resolveQueriesArgs[1].queries.length, 1);
-    t.regex(resolveQueriesArgs[1].queries[0], /.*\/additional1$/);
+    t.true(resolveQueriesArgs[1].queries[0].endsWith(`${path.sep}additional1`));
     t.deepEqual(resolveQueriesArgs[2].queries.length, 1);
-    t.regex(resolveQueriesArgs[2].queries[0], /.*\/additional2$/);
+    t.true(resolveQueriesArgs[2].queries[0].endsWith(`${path.sep}additional2`));
     t.deepEqual(resolveQueriesArgs[3].queries.length, 1);
-    t.regex(resolveQueriesArgs[3].queries[0], /.*\/foo$/);
+    t.true(resolveQueriesArgs[3].queries[0].endsWith(`${path.sep}foo`));
 
     // Now check that the end result contains all the queries
     t.deepEqual(config.queries["javascript"].builtin.length, 1);
     t.deepEqual(config.queries["javascript"].custom.length, 3);
-    t.regex(
-      config.queries["javascript"].builtin[0],
-      /javascript-code-scanning.qls$/
+    t.true(
+      config.queries["javascript"].builtin[0].endsWith(
+        "javascript-code-scanning.qls"
+      )
     );
-    t.regex(
-      config.queries["javascript"].custom[0].queries[0],
-      /.*\/additional1$/
+    t.true(
+      config.queries["javascript"].custom[0].queries[0].endsWith(
+        `${path.sep}additional1`
+      )
     );
-    t.regex(
-      config.queries["javascript"].custom[1].queries[0],
-      /.*\/additional2$/
+    t.true(
+      config.queries["javascript"].custom[1].queries[0].endsWith(
+        `${path.sep}additional2`
+      )
     );
-    t.regex(config.queries["javascript"].custom[2].queries[0], /.*\/foo$/);
+    t.true(
+      config.queries["javascript"].custom[2].queries[0].endsWith(
+        `${path.sep}foo`
+      )
+    );
   });
 });
 
@@ -859,10 +882,10 @@ test("Invalid queries in workflow file handled correctly", async (t) => {
         undefined,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         codeQL,
         tmpDir,
@@ -927,10 +950,10 @@ test("API client used when reading remote config", async (t) => {
       configFile,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -957,10 +980,10 @@ test("Remote config handles the case where a directory is provided", async (t) =
         repoReference,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -995,10 +1018,10 @@ test("Invalid format of remote config handled correctly", async (t) => {
         repoReference,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -1034,10 +1057,10 @@ test("No detected languages", async (t) => {
         undefined,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         codeQL,
         tmpDir,
@@ -1065,10 +1088,10 @@ test("Unknown languages", async (t) => {
         undefined,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         getCachedCodeQL(),
         tmpDir,
@@ -1118,10 +1141,10 @@ test("Config specifies packages", async (t) => {
       configFile,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example " },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -1175,10 +1198,10 @@ test("Config specifies packages for multiple languages", async (t) => {
       configFile,
       undefined,
       false,
+      false,
       "",
       "",
       { owner: "github", repo: "example" },
-      tmpDir,
       tmpDir,
       codeQL,
       tmpDir,
@@ -1243,10 +1266,10 @@ function doInvalidInputTest(
           configFile,
           undefined,
           false,
+          false,
           "",
           "",
           { owner: "github", repo: "example " },
-          tmpDir,
           tmpDir,
           codeQL,
           tmpDir,
@@ -1820,10 +1843,10 @@ const mlPoweredQueriesMacro = test.macro({
         undefined,
         undefined,
         false,
+        false,
         "",
         "",
         { owner: "github", repo: "example " },
-        tmpDir,
         tmpDir,
         codeQL,
         tmpDir,