+++ src/feature-flags.ts
@@ -10,6 +10,7 @@ export interface FeatureFlags {
 export enum FeatureFlag {
   MlPoweredQueriesEnabled = "ml_powered_queries_enabled",
   LuaTracerConfigEnabled = "lua_tracer_config_enabled",
+  TrapCachingEnabled = "trap_caching_enabled",
 }
 
 /**
@@ -31,14 +32,21 @@ export class GitHubFeatureFlags implements FeatureFlags {
   ) {}
 
   async getValue(flag: FeatureFlag): Promise<boolean> {
-    const response = (await this.getApiResponse())[flag];
+    const response = await this.getApiResponse();
     if (response === undefined) {
+      this.logger.debug(
+        `No feature flags API response for ${flag}, considering it disabled.`
+      );
+      return false;
+    }
+    const flagValue = response[flag];
+    if (flagValue === undefined) {
       this.logger.debug(
         `Feature flag '${flag}' undefined in API response, considering it disabled.`
       );
       return false;
     }
-    return response;
+    return flagValue;
   }
 
   private async getApiResponse(): Promise<FeatureFlagsApiResponse> {