+++ src/util.test.ts
@@ -335,7 +335,6 @@ for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
         pathsIgnore: [],
         originalUserInput: {},
         tempDir: tmpDir,
-        toolCacheDir: tmpDir,
         codeQLCmd: "",
         gitHubVersion: {
           type: util.GitHubVariant.DOTCOM,
@@ -348,6 +347,7 @@ for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
         debugArtifactName: util.DEFAULT_DEBUG_ARTIFACT_NAME,
         debugDatabaseName: util.DEFAULT_DEBUG_DATABASE_NAME,
         injectedMlQueries: false,
+        trapCaches: {},
       };
 
       t.is(util.getMlPoweredJsQueriesStatus(config), expectedStatus);